api/requirements.txt
aniso8601==8.0.0
attrs
python-dotenv
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.14.0
Click==7.0
coloredlogs==10.0
dnspython==1.16.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
flask-restplus==0.10.1
gunicorn==19.9.0
humanfriendly==4.18
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
marshmallow==3.2.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
pyrsistent==0.15.4
pytest==6.2.2
pytz==2019.2
PyYAML==5.3
six==1.12.0
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==0.16.0
zipp==3.4.0
